+++ src/utils.c
@@ -223,8 +223,7 @@ void truncate_if_needed(const char* source, int max_src_len, char* dest, int max
             return;
         }
 
-        if ((size_t) truncate_index > max_dest_len ||
-            truncate_index + truncate_str_len > max_dest_len ||
+        if (truncate_index > max_dest_len || truncate_index + truncate_str_len > max_dest_len ||
             index_after_replacement + remaining_chars_from_src > max_dest_len ||
             remaining_source_index + remaining_chars_from_src > max_src_len) {
             return;
